Remove long

+++ _printf.c
@@ -8,7 +8,7 @@ int _printf(const char *format, ...)
 {
 	const char *p;
 	unsigned int i;
-	long int j;
+	int j;
 	int k = 0;
 	va_list conspec;
 	char *s;
@@ -40,11 +40,11 @@ int _printf(const char *format, ...)
 			_putchar('%', &k);
 			break;
 		case 'd':
-			j = va_arg(conspec, long int);
+			j = va_arg(conspec, int);
 			_print_number(j, &k);
 			break;
 		case 'i':
-			j = va_arg(conspec, long int);
+			j = va_arg(conspec, int);
 			_print_number(j, &k);
 			break;
 		case 'r':
